add lift 2s and ac1
